requirements.txt
openai
pyttsx3
google-cloud-texttospeech
google-cloud-speech
pydub
sounddevice
scipy
wavio
torch>=1.0.2
torchaudio >= 0.9.0
